fix: moved models to respective pkgs
