+++ event/model.go
@@ -1,18 +1,10 @@
-package database
+package event
 
 import (
 	"time"
 
 	"github.com/google/uuid"
 )
-
-type User struct {
-	ID        uuid.UUID `db:"id"`
-	Name      string    `db:"name"`
-	Email     string    `db:"email"`
-	Password  string    `db:"password"`
-	CreatedAt time.Time `db:"created_at"`
-}
 
 type Event struct {
 	ID        uuid.UUID `db:"id"`
